Retain 1.6 on tomcat 8 for now
Travis-ci jdk 6 build will fail unless we do this.  While eclipse shows
compliancy issues it was working even on the early builds since the code
itself was not at the right level.  This does not prevent us from
running all the way back to java 4 on tomcat 5.0.x.  It just won't load
these which is fine.