[tests] Skip Date fields on javabean-tester as it fails under certain conditions
It's not deterministic when this will fail.  It grabs a constructor and
happens to get string to use.  The string is not parsed by date and
fails.  Because of multithreading, this occasionally was working.  It's
more of a problem now as all data objects are allowed to call real
constructors (first occurrence) rather than default no-arg constructor.
Hence a bug in javabean-tester that will now need addressed.