[ci] Patch for classNotFound in initialContext lookup
Example of this condition is using CDI wired in at application
classloader.  Tomcat cannot see it in how we are trying to access it and
therefore we will fail to find the class.  As such, just run as a
warning for now instead of throwing a stack trace with error.