[secure] Per nvs, add htmlEscape to param usage
param.webapp is the name of the webapp.  This comes from server not client but nvs cannot figure that out.  Unless user can inject into parameters which is already blocked, this should not be a problem.  However, to ensure we make nvs trust our usage, this causes no degrade in how things work.