[tomcat10] Drop notion internally of difference between 10 and 10.1
due to our spring usage, we could have never went here anyways.  Since 10/31 marked 10.0 as end of life, just focus on 10.1 now reducing our modules.  Will have to sync code under before deleting the extra module.  This just sets it so in code when we get to this point later.