[remove] Remove obsolete bonecp
bonecp has not seen a release since 2013 and has been flagged deprecated since 2014 with suggestion to use HikariCP instead.  while openejb still is supporting it, they are based off the older 2011 release.  see no reason to continue to support such long dead project at this point.  with upcoming migration to jakarta its also possible to encounter issues as it is still on javax namespace although not confirmed how much of that is still on java end.