[openejb] Fix to be on correct version as 8.0.15
after writing a test and realizing it was mixing core dbcp, it seemed like something was wrong and of course it was.  openejb 7 is equiv to tomcat 8.5.  8 is equiv to 9.  In our core we run latest as provided and both work same for our needs after fixing up and removing global exclusion it had.