[ojdbc] Port code back from psi-probe at commit 40b76f7306cbb691560b1093fb18a4f7fbfe39f1
was separated originally due to special concerns of working with oracle drivers.  They opened sourced this now many years back so the overhead of keeping it separate didn't make much sense especially when the interfaces used in psi probe had to live over there.  The old repo is now archived.