Remove deprecated spring theme and use css directly
while novel to possible support different theme, this never did support anything but the theme it had.  Spring has since deprecated and after looking, it was clear the theme was not used everywhere as it where.  So simply remove and use directly now to clear that deprecation while being compliant with older usage.